+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSetFactory.java
@@ -559,14 +559,17 @@ public class RuleSetFactory {
             if (rulesetDeprecated || !r.getRule().isDeprecated()) {
                 // add the rule, if either the ruleset itself is deprecated (then we add all rules)
                 // or if the rule is not deprecated (in that case, the ruleset might contain deprecated as well
-                // as valid references to rules)
+                // as valid rules)
                 ruleSetBuilder.addRuleIfNotExists(r);
             }
         }
 
         if (!excludedRulesCheck.isEmpty()) {
-            throw new IllegalArgumentException(
-                    "Unable to exclude rules " + excludedRulesCheck + "; perhaps the rule name is mispelled?");
+            if (LOG.isLoggable(Level.WARNING)) {
+                LOG.warning(
+                    "Unable to exclude rules " + excludedRulesCheck + " from ruleset reference " + ref
+                    + "; perhaps the rule name is mispelled or the rule doesn't exist anymore?");
+            }
         }
     }
 
@@ -628,7 +631,7 @@ public class RuleSetFactory {
 
         // load the ruleset with minimum priority low, so that we get all rules, to be able to exclude any rule
         // minimum priority will be applied again, before constructing the final ruleset
-        RuleSetFactory ruleSetFactory = new RuleSetFactory(resourceLoader, RulePriority.LOW, warnDeprecated, this.compatibilityFilter != null);
+        RuleSetFactory ruleSetFactory = new RuleSetFactory(resourceLoader, RulePriority.LOW, false, this.compatibilityFilter != null);
 
         boolean isSameRuleSet = false;
         RuleSetReferenceId otherRuleSetReferenceId = RuleSetReferenceId.parse(ref).get(0);
@@ -676,7 +679,7 @@ public class RuleSetFactory {
 
         RuleReference ruleReference = new RuleFactory(resourceLoader).decorateRule(referencedRule, ruleSetReference, ruleElement);
 
-        if (warnDeprecated && ruleReference.isDeprecated()) {
+        if (warnDeprecated && ruleReference.isDeprecated() && !isSameRuleSet) {
             if (LOG.isLoggable(Level.WARNING)) {
                 LOG.warning("Use Rule name " + ruleReference.getRuleSetReference().getRuleSetFileName() + '/'
                         + ruleReference.getOriginalName() + " instead of the deprecated Rule name "